Refactoring the code into smaller components
